Refactor FTP services into plugin projects
