+++ DesktopApplicationTemplate.Service/DesktopApplicationTemplate.Service.csproj
@@ -14,9 +14,6 @@
     <PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.Logging.EventLog" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="8.0.0" />
-    <PackageReference Include="FluentFTP" Version="53.0.1" />
-    <PackageReference Include="FubarDev.FtpServer" Version="3.1.2" />
-    <PackageReference Include="FubarDev.FtpServer.FileSystem.DotNet" Version="3.1.2" />
   </ItemGroup>
 
   <ItemGroup>
