+++ DesktopApplicationTemplate.UI/DesktopApplicationTemplate.UI.csproj
@@ -14,7 +14,6 @@
 	</PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="FluentFTP" Version="53.0.1" />
     <PackageReference Include="SSH.NET" Version="2025.0.0" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="4.12.0" />
     <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.12.0" />
