--- DesktopApplicationTemplate.UI/Views/Ftp/Advanced/FtpServerAdvancedConfigView.xaml
@@ -1,35 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Ftp.Advanced.FtpServerAdvancedConfigView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:views="clr-namespace:DesktopApplicationTemplate.UI.Views"
-      mc:Ignorable="d">
-    <Grid Margin="20">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="Auto" />
-            <ColumnDefinition Width="*" />
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" /> <!-- Allow Anonymous -->
-            <RowDefinition Height="Auto" /> <!-- Username -->
-            <RowDefinition Height="Auto" /> <!-- Password -->
-            <RowDefinition Height="Auto" /> <!-- Buttons -->
-        </Grid.RowDefinitions>
-
-        <CheckBox Grid.Row="0" Grid.ColumnSpan="2" Content="Allow Anonymous" IsChecked="{Binding AllowAnonymous}" Style="{StaticResource FormField}"/>
-
-        <TextBlock Grid.Row="1" Grid.Column="0" Text="Username" Style="{StaticResource FormLabel}"/>
-        <TextBox Grid.Row="1" Grid.Column="1" Text="{Binding Username, UpdateSourceTrigger=PropertyChanged}" Style="{StaticResource FormField}"/>
-
-        <TextBlock Grid.Row="2" Grid.Column="0" Text="Password" Style="{StaticResource FormLabel}"/>
-        <TextBox Grid.Row="2" Grid.Column="1" Text="{Binding Password, UpdateSourceTrigger=PropertyChanged}" Style="{StaticResource FormField}"/>
-
-        <views:AdvancedConfigButtonBar Grid.Row="3" Grid.ColumnSpan="2"
-                                       SaveCommand="{Binding SaveCommand}"
-                                       BackCommand="{Binding BackCommand}"
-                                       SaveButtonText="Save Configuration"
-                                       SaveAutomationName="Save FTP Advanced Options"
-                                       BackAutomationName="Back from FTP Advanced Options"/>
-    </Grid>
-</Page>
