--- DesktopApplicationTemplate.UI/Views/Ftp/Advanced/FtpServerAdvancedConfigView.xaml.cs
@@ -1,26 +1,0 @@
-using System;
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels.Ftp.Advanced;
-
-namespace DesktopApplicationTemplate.UI.Views.Ftp.Advanced;
-
-public partial class FtpServerAdvancedConfigView : Page
-{
-    private readonly ILoggingService _logger;
-
-    public FtpServerAdvancedConfigView(ILoggingService logger)
-    {
-        InitializeComponent();
-        _logger = logger;
-    }
-
-    public void Initialize(FtpServerAdvancedConfigViewModel vm)
-    {
-        if (vm is null)
-            throw new ArgumentNullException(nameof(vm));
-
-        DataContext = vm;
-        vm.Logger = _logger;
-    }
-}
