--- DesktopApplicationTemplate.UI/Views/Ftp/Create/FtpServerCreateView.xaml.cs
@@ -1,15 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels.Ftp.Create;
-
-namespace DesktopApplicationTemplate.UI.Views.Ftp.Create;
-
-public partial class FtpServerCreateView : Page
-{
-    public FtpServerCreateView(FtpServerCreateViewModel vm, ILoggingService logger)
-    {
-        InitializeComponent();
-        DataContext = vm;
-        vm.Logger = logger;
-    }
-}
