--- DesktopApplicationTemplate.UI/Views/Ftp/Edit/FtpServerEditView.xaml
@@ -1,39 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Ftp.Edit.FtpServerEditView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:views="clr-namespace:DesktopApplicationTemplate.UI.Views"
-      mc:Ignorable="d">
-    <ScrollViewer VerticalScrollBarVisibility="Auto">
-        <Grid Margin="20">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" /> <!-- Service Name -->
-                <RowDefinition Height="Auto" /> <!-- Port -->
-                <RowDefinition Height="Auto" /> <!-- Root Path -->
-                <RowDefinition Height="Auto" /> <!-- Buttons -->
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0" Grid.Column="0" Text="Service Name" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="0" Grid.Column="1" Text="{Binding ServiceName, UpdateSourceTrigger=PropertyChanged}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="1" Grid.Column="0" Text="Port" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="1" Grid.Column="1" Text="{Binding Port, UpdateSourceTrigger=PropertyChanged}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="2" Grid.Column="0" Text="Root Path" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="2" Grid.Column="1" Text="{Binding RootPath, UpdateSourceTrigger=PropertyChanged}" Style="{StaticResource FormField}"/>
-
-            <views:EditorButtonBar Grid.Row="3" Grid.ColumnSpan="2"
-                                   AdvancedConfigCommand="{Binding AdvancedConfigCommand}"
-                                   SaveCommand="{Binding SaveCommand}"
-                                   CancelCommand="{Binding CancelCommand}"
-                                   AdvancedAutomationName="Open FTP Advanced Configuration"
-                                   SaveAutomationName="Save FTP Server"
-                                   CancelAutomationName="Cancel FTP Server Edit" />
-        </Grid>
-    </ScrollViewer>
-</Page>
