--- DesktopApplicationTemplate.UI/Views/Ftp/Edit/FtpServerEditView.xaml.cs
@@ -1,17 +1,0 @@
-using System;
-using System.Windows.Controls;
-using DesktopApplicationTemplate.UI.ViewModels.Ftp.Edit;
-
-namespace DesktopApplicationTemplate.UI.Views.Ftp.Edit;
-
-/// <summary>
-/// Interaction logic for FtpServerEditView.
-/// </summary>
-public partial class FtpServerEditView : Page
-{
-    public FtpServerEditView(FtpServerEditViewModel viewModel)
-    {
-        DataContext = viewModel ?? throw new ArgumentNullException(nameof(viewModel));
-        InitializeComponent();
-    }
-}
