--- DesktopApplicationTemplate.UI/Views/Ftp/FTPServiceView.xaml
@@ -1,94 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Ftp.FTPServiceView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:helpers="clr-namespace:DesktopApplicationTemplate.UI.Helpers"
-      xmlns:shared="clr-namespace:DesktopApplicationTemplate.UI.Views.Shared"
-      mc:Ignorable="d">
-    <Page.Resources>
-        <helpers:BooleanToBrushConverter x:Key="BooleanToBrushConverter" TrueBrush="Green" FalseBrush="Red"/>
-        <helpers:StringNullOrEmptyToVisibilityConverter x:Key="StringNullOrEmptyToVisibilityConverter" />
-        <helpers:NullToVisibilityConverter x:Key="NullToVisibilityConverter" />
-    </Page.Resources>
-    <Grid Margin="10">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-        <Border Grid.Row="0" CornerRadius="8" BorderThickness="1" Padding="10" Margin="0,0,0,10"
-                Background="{Binding Tag.BackgroundColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                BorderBrush="{Binding Tag.BorderColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                ToolTip="{Binding Tag.DescriptorTooltip, RelativeSource={RelativeSource AncestorType=Page}}">
-            <DockPanel>
-                <Border Width="40" Height="40" CornerRadius="20" Margin="0,0,10,0"
-                        Background="{Binding Tag.BorderColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                        Visibility="{Binding Tag.IconGlyph, RelativeSource={RelativeSource AncestorType=Page}, Converter={StaticResource NullToVisibilityConverter}}">
-                    <TextBlock Text="{Binding Tag.IconGlyph, RelativeSource={RelativeSource AncestorType=Page}}"
-                               FontSize="22" HorizontalAlignment="Center" VerticalAlignment="Center" />
-                </Border>
-                <StackPanel>
-                    <TextBlock Text="{Binding Tag.DisplayName, RelativeSource={RelativeSource AncestorType=Page}}" FontWeight="Bold" />
-                    <TextBlock Text="{Binding Tag.DescriptorDescription, RelativeSource={RelativeSource AncestorType=Page}}"
-                               Foreground="Gray" TextWrapping="Wrap"
-                               Visibility="{Binding Tag.DescriptorDescription, RelativeSource={RelativeSource AncestorType=Page}, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}" />
-                </StackPanel>
-            </DockPanel>
-        </Border>
-        <StackPanel Grid.Row="1" Orientation="Horizontal">
-            <Ellipse Width="10" Height="10" Fill="{Binding IsServerRunning, Converter={StaticResource BooleanToBrushConverter}}" Margin="0,0,5,0"/>
-            <TextBlock Text="{Binding ServerStatus}" FontWeight="Bold" FontSize="14"/>
-            <TextBlock Text="{Binding ConnectedClients, StringFormat=Clients: {0}}" Margin="10,0,0,0"/>
-        </StackPanel>
-        <StackPanel Grid.Row="2" Orientation="Horizontal" Margin="0,10,0,0">
-            <Button Content="Start" Width="100" Margin="0,0,5,0" Command="{Binding StartServerCommand}"/>
-            <Button Content="Stop" Width="100" Command="{Binding StopServerCommand}"/>
-        </StackPanel>
-        <GroupBox Header="Active Transfers" Grid.Row="3" Margin="0,10,0,0">
-            <ItemsControl ItemsSource="{Binding Transfers}">
-                <ItemsControl.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel Margin="0,0,0,5">
-                            <TextBlock Text="{Binding Path}"/>
-                            <ProgressBar Minimum="0" Maximum="1" Value="{Binding Progress}" Height="10"/>
-                        </StackPanel>
-                    </DataTemplate>
-                </ItemsControl.ItemTemplate>
-            </ItemsControl>
-        </GroupBox>
-        <GroupBox Header="Uploaded Files" Grid.Row="4" Margin="0,10,0,0">
-            <ListBox ItemsSource="{Binding UploadedFiles}">
-                <ListBox.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel Orientation="Horizontal">
-                            <TextBlock Text="{Binding Path}"/>
-                            <TextBlock Text=" ("/>
-                            <TextBlock Text="{Binding Size}"/>
-                            <TextBlock Text=" bytes)"/>
-                        </StackPanel>
-                    </DataTemplate>
-                </ListBox.ItemTemplate>
-            </ListBox>
-        </GroupBox>
-        <GroupBox Header="Downloaded Files" Grid.Row="5" Margin="0,10,0,0">
-            <ListBox ItemsSource="{Binding DownloadedFiles}">
-                <ListBox.ItemTemplate>
-                    <DataTemplate>
-                        <StackPanel Orientation="Horizontal">
-                            <TextBlock Text="{Binding Path}"/>
-                            <TextBlock Text=" ("/>
-                            <TextBlock Text="{Binding Size}"/>
-                            <TextBlock Text=" bytes)"/>
-                        </StackPanel>
-                    </DataTemplate>
-                </ListBox.ItemTemplate>
-            </ListBox>
-        </GroupBox>
-        <shared:ServiceLogView x:Name="LogView" Grid.Row="6" Margin="0,10,0,0" />
-    </Grid>
-</Page>
