+++ DesktopApplicationTemplate.sln
@@ -32,6 +32,11 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HttpUI", "Services\Http\HttpUI\HttpUI.csproj", "{A8B98B13-DA95-43BA-B2F4-8604368818E7}"
 EndProject
 
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FtpPackage", "Services\Ftp\FtpPackage\FtpPackage.csproj", "{6A9F33F7-E8BF-416B-B37C-8D0B384E47D2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FtpUI", "Services\Ftp\FtpUI\FtpUI.csproj", "{389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}"
+EndProject
+
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -94,6 +99,14 @@ Global
                 {9D22FB33-5081-418F-961A-20CFEFD1C584}.Debug|Any CPU.Build.0 = Debug|Any CPU
                 {9D22FB33-5081-418F-961A-20CFEFD1C584}.Release|Any CPU.ActiveCfg = Release|Any CPU
                 {9D22FB33-5081-418F-961A-20CFEFD1C584}.Release|Any CPU.Build.0 = Release|Any CPU
+                {6A9F33F7-E8BF-416B-B37C-8D0B384E47D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {6A9F33F7-E8BF-416B-B37C-8D0B384E47D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {6A9F33F7-E8BF-416B-B37C-8D0B384E47D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {6A9F33F7-E8BF-416B-B37C-8D0B384E47D2}.Release|Any CPU.Build.0 = Release|Any CPU
+                {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {389FAA88-2E7D-4093-ABB1-F6AAB6D15E38}.Release|Any CPU.Build.0 = Release|Any CPU
         EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
