+++ Services/Ftp/FtpPackage/FtpPackage.csproj
@@ -1,0 +1,28 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <Import Project="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.props" Condition="Exists('..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.props')" />
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <AssemblyName>DesktopApplicationTemplate.Ftp</AssemblyName>
+    <RootNamespace>DesktopApplicationTemplate.Services.Ftp</RootNamespace>
+    <ServicePluginPackageId>DesktopApplicationTemplate.Ftp</ServicePluginPackageId>
+    <ServicePluginPackageVersion>1.0.0</ServicePluginPackageVersion>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\DesktopApplicationTemplate.Core\DesktopApplicationTemplate.Core.csproj" />
+    <ProjectReference Include="..\..\..\DesktopApplicationTemplate.Services.Common\DesktopApplicationTemplate.Services.Common.csproj" />
+    <ProjectReference Include="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.csproj" ReferenceOutputAssembly="false" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="FluentFTP" Version="53.0.1" />
+    <PackageReference Include="FubarDev.FtpServer" Version="3.1.2" />
+    <PackageReference Include="FubarDev.FtpServer.FileSystem.DotNet" Version="3.1.2" />
+    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
+  </ItemGroup>
+
+  <Import Project="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.targets" Condition="Exists('..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.targets')" />
+</Project>
