+++ Services/Ftp/FtpPackage/plugin.manifest.json
@@ -1,0 +1,12 @@
+{
+  "id": "DesktopApplicationTemplate.Ftp",
+  "name": "FTP Service Package",
+  "version": "1.0.0",
+  "entryAssembly": "DesktopApplicationTemplate.Ftp.dll",
+  "serviceAssemblies": [
+    "DesktopApplicationTemplate.Ftp.dll"
+  ],
+  "probingPaths": [
+    "libs"
+  ]
+}
