+++ Services/Ftp/FtpUI/Views/Ftp/Advanced/FtpServerAdvancedConfigView.xaml.cs
@@ -1,0 +1,26 @@
+using System;
+using System.Windows.Controls;
+using DesktopApplicationTemplate.Core.Services;
+using DesktopApplicationTemplate.Services.Ftp.UI.ViewModels.Ftp.Advanced;
+
+namespace DesktopApplicationTemplate.Services.Ftp.UI.Views.Ftp.Advanced;
+
+public partial class FtpServerAdvancedConfigView : Page
+{
+    private readonly ILoggingService _logger;
+
+    public FtpServerAdvancedConfigView(ILoggingService logger)
+    {
+        InitializeComponent();
+        _logger = logger;
+    }
+
+    public void Initialize(FtpServerAdvancedConfigViewModel vm)
+    {
+        if (vm is null)
+            throw new ArgumentNullException(nameof(vm));
+
+        DataContext = vm;
+        vm.Logger = _logger;
+    }
+}
