+++ Services/Ftp/FtpUI/plugin.manifest.json
@@ -1,0 +1,12 @@
+{
+  "id": "DesktopApplicationTemplate.Ftp.UI",
+  "name": "FTP Service UI",
+  "version": "1.0.0",
+  "entryAssembly": "DesktopApplicationTemplate.Ftp.UI.dll",
+  "serviceAssemblies": [
+    "DesktopApplicationTemplate.Ftp.UI.dll"
+  ],
+  "probingPaths": [
+    "libs"
+  ]
+}
